# contracts/cairn-openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CreateGrantRequest": {
        "additionalProperties": false,
        "properties": {
          "grant": {
            "$ref": "#/components/schemas/GrantBody"
          },
          "realm_id": {
            "title": "Realm Id",
            "type": "string"
          }
        },
        "required": [
          "realm_id",
          "grant"
        ],
        "title": "CreateGrantRequest",
        "type": "object"
      },
...
      "CreateGrantResultEnvelope": {
        "additionalProperties": false,
        "properties": {
          "audit_receipt": {
            "$ref": "#/components/schemas/AuditReceiptBody"
          },
          "mutation_receipt": {
            "$ref": "#/components/schemas/MutationReceiptBody"
          },
          "outcome": {
            "$ref": "#/components/schemas/WireOutcome"
          },
          "result": {
            "$ref": "#/components/schemas/CreateGrantResult"
          }
        },
        "required": [
          "outcome",
          "result",
          "mutation_receipt",
          "audit_receipt"
        ],
        "title": "SuccessEnvelope[CreateGrantResult]",
        "type": "object"
      },
...
      "FailureEnvelope": {
        "additionalProperties": false,
        "properties": {
          "failure": {
            "$ref": "#/components/schemas/FailureBody"
          }
        },
        "required": [
          "failure"
        ],
        "title": "FailureEnvelope",
        "type": "object"
      },
...
      "InstanceResult": {
        "additionalProperties": false,
        "properties": {
          "contract_digest": {
            "title": "Contract Digest",
            "type": "string"
          },
          "contract_identity": {
            "title": "Contract Identity",
            "type": "string"
          },
          "instance_id": {
            "title": "Instance Id",
            "type": "string"
          },
          "mcp_contract_digest": {
            "title": "Mcp Contract Digest",
            "type": "string"
          },
          "product_version": {
            "title": "Product Version",
            "type": "string"
          }
        },
        "required": [
          "instance_id",
          "product_version",
          "contract_identity",
          "contract_digest"
        ],
        "title": "InstanceResult",
        "type": "object"
      },
...
      "MutationReceiptBody": {
        "additionalProperties": false,
        "properties": {
          "command_digest": {
            "title": "Command Digest",
            "type": "string"
          },
          "mutation_id": {
            "title": "Mutation Id",
            "type": "string"
          }
        },
        "required": [
          "mutation_id",
          "command_digest"
        ],
        "title": "MutationReceiptBody",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "An opaque Cairn credential. TLS is mandatory outside loopback.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The Cairn v0.1 memory authority surface. Every operation authenticates with an opaque bearer credential; scope isolation is enforced server-side and sibling scopes are never visible. Route names are the closed operation vocabulary itself, so this document is a one-to-one map of the authority model.\n\nTwo boundary behaviours are deliberately absent from the per-operation responses below, because neither is a response to a valid call on the operation it would sit under: an unknown path returns the fixed not_found failure, and a known path with a wrong method returns invalid_request with status 405.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "summary": "The cairn/v1 contract.",
    "title": "Cairn",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/create-grant": {
      "post": {
        "operationId": "create_grant",
        "parameters": [
          {
            "description": "A caller-supplied UUIDv4 is adopted; any other value is ignored in favour of a fresh one, never rejected. Echoed on every response.",
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "A canonical lowercase hyphenated UUID. Replaying one returns the original receipt with outcome 'replayed'.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateGrantResultEnvelope"
                }
              }
            },
            "description": "The operation's result.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: invalid_request, secret_rejected.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: authentication_failed.",
            "headers": {
              "WWW-Authenticate": {
                "description": "The bearer challenge.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: authorisation_denied.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: not_found.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: idempotency_conflict.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: invalid_request.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: invalid_request.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "421": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: instance_mismatch.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: internal_error.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: dependency_unavailable, index_pending, stale_index.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Create a grant over a scope prefix."
      }
    },
...
    "/v1/instance": {
      "get": {
        "description": "This is a read: supplying an Idempotency-Key header is invalid_request. The header is required on all eight mutations and refused on both reads.",
        "operationId": "instance",
        "parameters": [
          {
            "description": "A caller-supplied UUIDv4 is adopted; any other value is ignored in favour of a fresh one, never rejected. Echoed on every response.",
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceResult"
                }
              }
            },
            "description": "The operation's result.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: invalid_request, secret_rejected.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: authentication_failed.",
            "headers": {
              "WWW-Authenticate": {
                "description": "The bearer challenge.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: authorisation_denied.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: not_found.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: idempotency_conflict.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "421": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: instance_mismatch.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: internal_error.",
            "headers": {
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureEnvelope"
                }
              }
            },
            "description": "Failure: dependency_unavailable, index_pending, stale_index.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-Correlation-ID": {
                "description": "The correlation identifier for this request.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Report the contract identity, product version and contract digest."
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}